+++ NEWS.md
@@ -1,3 +1,5 @@
+# pkgcache (development version)
+
 # pkgcache 2.2.2
 
 * pkgcache now treats R 4.5.x (current R-devel) macOS binaries